# docs/api-docs.json
{"openapi":"3.0.3","info":{"title":"Api Documentation","description":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0"},"servers":[{"url":"http://localhost:8082","description":"Inferred Url"}],"tags":[{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"building-controller","description":"Building Controller"},{"name":"resource-controller","description":"Resource Controller"}],"paths":{"/error":{"get":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPUT","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPOST","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingDELETE","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingOPTIONS","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"head":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingHEAD","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPATCH","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"trace":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingTRACE","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/building":{"get":{"tags":["building-controller"],"summary":"Returns all registered buildings","operationId":"getAllBuildingsUsingGET","responses":{"200":{"description":"Returns a list of buildings","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BuildingDTO"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["building-controller"],"summary":"Create a building","operationId":"createBuildingUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}}}},"responses":{"200":{"description":"Returns a building","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}},"*/*":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}}}}},"/building/{building_id}":{"get":{"tags":["building-controller"],"summary":"Returns a registered building that has the informed ID","operationId":"getBuildingUsingGET","parameters":[{"name":"buildingId","in":"path","description":"buildingId","required":true,"style":"simple","schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Returns a building","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}},"*/*":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}}}},"put":{"tags":["building-controller"],"summary":"Put a building","operationId":"putBuildingUsingPUT","parameters":[{"name":"buildingId","in":"path","description":"buildingId","required":true,"style":"simple","schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}}}},"201":{"description":"Created"},"204":{"description":"Returns a building","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}}}},"delete":{"tags":["building-controller"],"summary":"Delete a building","operationId":"deleteBuildingUsingDELETE","parameters":[{"name":"buildingId","in":"path","description":"buildingId","required":true,"style":"simple","schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}},"204":{"description":"Returns a building","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}}}},"patch":{"tags":["building-controller"],"summary":"Patch a building","operationId":"patchBuildingUsingPATCH","parameters":[{"name":"buildingId","in":"path","description":"buildingId","required":true,"style":"simple","schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}}}},"204":{"description":"Returns a building","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingDTO"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}}}}},"/resource":{"get":{"tags":["resource-controller"],"summary":"Returns all registered resources","operationId":"getAllResourcesUsingGET","responses":{"200":{"description":"Returns a list of resources","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResourceDTO"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["resource-controller"],"summary":"Create a resource","operationId":"createResourceUsingPOST","parameters":[{"name":"caracteristics[0].id","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"caracteristics[0].name","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"description","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"id","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"resourceType.id","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"resourceType.name","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"status","in":"query","required":false,"style":"form","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a resource","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDTO"}},"*/*":{"schema":{"$ref":"#/components/schemas/ResourceDTO"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}}}}},"/resource/{building_id}":{"put":{"tags":["resource-controller"],"summary":"Put a resource","operationId":"putResourceUsingPUT","parameters":[{"name":"caracteristics[0].id","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"caracteristics[0].name","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"description","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"id","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"resourceId","in":"query","description":"resourceId","required":true,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"resourceType.id","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"resourceType.name","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"status","in":"query","required":false,"style":"form","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResourceDTO"}}}},"201":{"description":"Created"},"204":{"description":"Returns a resource","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDTO"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}}}}},"/resource/{resource_id}":{"get":{"tags":["resource-controller"],"summary":"Returns a registered resource that has the informed ID","operationId":"getResourceUsingGET","parameters":[{"name":"resourceId","in":"query","description":"resourceId","required":true,"style":"form","schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Returns a resource","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDTO"}},"*/*":{"schema":{"$ref":"#/components/schemas/ResourceDTO"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}}}},"delete":{"tags":["resource-controller"],"summary":"Delete a resource","operationId":"deleteResourceUsingDELETE","parameters":[{"name":"resourceId","in":"query","description":"resourceId","required":true,"style":"form","schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}},"204":{"description":"Returns a resource","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDTO"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}}}},"patch":{"tags":["resource-controller"],"summary":"Patch a resource","operationId":"patchResourceUsingPATCH","parameters":[{"name":"caracteristics[0].id","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"caracteristics[0].name","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"description","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"id","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"resourceId","in":"query","description":"resourceId","required":true,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"resourceType.id","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"resourceType.name","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"status","in":"query","required":false,"style":"form","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResourceDTO"}}}},"204":{"description":"Returns a resource","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDTO"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"Error-ModelName{namespace='com.rremiao.ngsarc.domain.dto', name='ExceptionDTO'}"}}}}}}}},"components":{"schemas":{"AddressDTO":{"title":"AddressDTO","type":"object","properties":{"cep":{"type":"string"},"complement":{"type":"string"},"district":{"type":"string"},"id":{"type":"integer","format":"int32"},"number":{"type":"integer","format":"int32"},"street":{"type":"string"},"uf":{"type":"string"}}},"BuildingDTO":{"title":"BuildingDTO","type":"object","properties":{"address":{"$ref":"#/components/schemas/AddressDTO"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"number":{"type":"integer","format":"int32"},"rooms":{"type":"array","items":{"$ref":"#/components/schemas/RoomDTO"}}}},"CaracteristicsDTO":{"title":"CaracteristicsDTO","type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"ModelAndView":{"title":"ModelAndView","type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/components/schemas/View"},"viewName":{"type":"string"}}},"ResourceDTO":{"title":"ResourceDTO","type":"object","properties":{"caracteristics":{"type":"array","items":{"$ref":"#/components/schemas/CaracteristicsDTO"}},"description":{"type":"string"},"id":{"type":"integer","format":"int32"},"resourceType":{"$ref":"#/components/schemas/ResourceTypeDTO"},"status":{"type":"string"}}},"ResourceTypeDTO":{"title":"ResourceTypeDTO","type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"RoomDTO":{"title":"RoomDTO","type":"object","properties":{"capacity":{"type":"integer","format":"int32"},"floor":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"number":{"type":"integer","format":"int32"}}},"View":{"title":"View","type":"object","properties":{"contentType":{"type":"string"}}}}}}