# src/main/java/com/rremiao/ngsarc/core/dto/BuildingDTO.java
package com.rremiao.ngsarc.core.dto;

import lombok.Data;
import java.util.List;

@Data
public class BuildingDTO {
    
    private int id;
    private String name;
    private int number;
    private AddressDTO address;
    private List<RoomDTO> rooms;
}